requirements.txt
accelerate==0.30.1
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
filelock==3.14.0
fsspec==2024.5.0
huggingface-hub==0.23.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
psutil==5.9.8
psycopg2-binary==2.9.9
python-decouple==3.8
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.1
safetensors==0.4.3
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
